Resync with S4Vectors 0.19.11
In S4Vectors 0.19.11 the default for the 'use.names' argument of the
mcols() getter was changed from FALSE to TRUE. Because of this change
the rowData() getter can be made an alias for mcols() again i.e. it
doesn't need to use dedicated code to raise an error in case of
duplicated rownames. However note that, like mcols(), it still fails
by default when the rownames contain NAs.