Use SimpleAssays instead of ShallowSimpleListAssays to store the assays
Wrap the user-supplied assays in a SimpleAssays object instead of a
ShallowSimpleListAssays object before putting them in the 'assays' slot
of a SummarizedExperiment object or derivative. Note that there are
probably hundreds (if not thousands) of serialized SummarizedExperiment
objects around that use ShallowSimpleListAssays. These objects should
keep working as before.

See https://github.com/Bioconductor/SummarizedExperiment/issues/21 for
the motivation behind this change.