By default assays()/assay() setters reject inconsistent dimnames
By default the dimnames on the supplied assay(s) must be identical to
the dimnames on the SummarizedExperiment object. The user now must use
'withDimnames=FALSE' if it's not the case or they get an error.
This is for symmetry with the behavior of the assays()/assay() getters
(see issue #35). Unfortunately this change is likely to break existing
code but at least the fix is easy.

Other changes:

- dimnames() now returns NULL instead of list(NULL, NULL) on a
  SummarizedExperiment object with no dimnames. This is consistent
  with matrix objects.

- Swap positions of arguments '...' and 'withDimnames' in assays()
  setter and getter. So now it's:

      assays(x, withDimnames=TRUE, ...)
      assays(x, withDimnames=TRUE, ...) <- value

- Add 'withDimnames' argument to the assay() getter/setter. So now it's:

      assay(x, i, withDimnames=TRUE, ...)
      assay(x, i, withDimnames=TRUE, ...) <- value

  Note that before this change, the user was able to explicitly set
  'withDimnames' when calling assay() but since this was not a formal
  argument it was forwarded to assays() via the ellipsis. Having it as
  a formal argument makes it easier to discover and allows tab completion.