Move MatrixGenerics from Imports to Depends
With MatrixGenerics in Imports, the following sequence:

    library(SummarizedExperiment)
    library(matrixStats)

breaks the rowRanges() getter for RangedSummarizedExperiment objects.
(Starting with SummarizedExperiment 1.19.8, this getter is a method for
the rowRanges() generic defined in MatrixGenerics.)

Having MatrixGenerics in Depends protects against this by ensuring that
MatrixGenerics ends up before matrixStats on the search path after loading
the SummarizedExperiment package.