Accept NAs in the rownames of a SummarizedExperiment object
IMPORTANT NOTES:
- NAs in the **rownames** are now tolerated but will cause problems
  downstream e.g. they break the rowData() getter unless 'use.names=FALSE'
  is used.
- NAs in the **colnames** are not and cannot be supported at the moment!
  Right now they break the SummarizedExperiment() constructor in an ugly
  way (error message not super helpful):
      > SummarizedExperiment(m)
      Error in DataFrame(x = seq_len(ncol(a1)), row.names = nms) :
        missing values in 'row.names'
  This should be improved.
- At the root of these problems is the fact that at the moment DataFrame
  objects do NOT support NAs in their rownames.

Bottom line: NAs in the dimnames of a SummarizedExperiment object should
be avoided at all cost. One way to deal with them is to replace them with
empty strings (""). It's not clear whether the SummarizedExperiment()
constructor should automatically take care of that (with a warning), or
if it should just fail with an informative error message asking the user
to "fix" the dimnames on the assay **before** calling the constructor.